+++ demo/src/main/java/com/orange/ods/demo/ui/components/cards/CardImageFirst.kt
@@ -24,7 +24,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import com.orange.ods.compose.component.card.OdsCardImageFirst
+import com.orange.ods.compose.component.card.OdsImageFirstCard
 import com.orange.ods.demo.R
 import com.orange.ods.demo.ui.components.utilities.ComponentCustomizationBottomSheetScaffold
 import com.orange.ods.demo.ui.components.utilities.clickOnElement
@@ -57,7 +57,7 @@ fun CardImageFirst() {
                 val button2Text = stringResource(id = R.string.component_element_button2)
                 val cardContainerText = stringResource(id = R.string.component_card_element_container)
 
-                OdsCardImageFirst(
+                OdsImageFirstCard(
                     title = stringResource(id = R.string.component_element_title),
                     image = painterResource(id = R.drawable.placeholder),
                     subtitle = if (hasSubtitle) stringResource(id = R.string.component_element_subtitle) else null,